refactor: Исправил все поставленные замечания по ТЗ.
